Eliminate unused `slice_name` accessor
The only consumers of the `slice_name` value, thankfully, were internal to the
slice code, and all used the member variable directly.

This eliminates the accessor allowing external read and write operations on
that (nominally constant) member, making it easier to understand that it is
purely internal data being tracked.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>