Automate the determination of the slice name
The manual management of the slice name really brought nothing to the table:
it was carefully duplicating data about the class name that was trivially
calculated by the machine.

This replaces all that careful manual data, and the member, with a simple
function that simply calculates and returns the correct value.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>