Don't pass slice name to the help printing helper
This stops passing the slice name to the common code that prints out the
command help on request - unsurprisingly, nobody ever invokes that for a slice
name other than the current slice name.

By simply asked for that in the helper, rather than asking for it externally
and passing it through, we can eliminate every reference to the slice name
outside the shared code in the base class.

This, again, simplifies the behaviour of the code, eliminates what looks like
state but really isn't, and makes it easier for new developers to approach
the codebase.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>