Eliminate unused `slice_commands` accessor on slice
The `slice_commands` data, which is modified by every slice, is never accessed
outside the class itself.  It is purely the internal dispatch map used to turn
command line input into command method execution.

This removes the unused public accessor for the data, making it clearer that
this isn't used by any other part of the code and firming up the API lines
between parts of the codebase.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>