Turn `slice_commands` into a private method, not instance data
The slice code used an instance variable (mutable, naturally) and object
initialization to ensure that the set of `slice_commands` could be
discovered correctly.

Aside from the unused mutability, this also made it harder to determine when
that data was used - the subclasses needed to be quite intimate with how the
parent class used that data.

This change moves that to a distinct method, and allows the normal work of
subclass overriding methods to free the subclass initializer from
understanding the parent class initialization flow vs command processing.

There is no functional change from this transformation, simply less mutable
state dragged around to confuse developers as they work on understanding how
this project operates.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>