Eliminate unhelpful default "help: n/a" help output...
The code that handled slice failures tried hard to display some sort of useful
help to the user, but included a default case where we would simply drop in
`n/a` to let the user know we couldn't find any help.

Sadly, the conditional there was such that this was *always* what displayed if
we didn't have hand-crafted custom help - it tried to access content in the
command hash that would never, ever exist.

Rather than including this quite unhelpful message, this eliminates the
persistent mut(il)ation of the command hash, the conditional, and the
unhelpful help message entirely.

The user still receives the exact same volume of help: if custom text is
written, that, otherwise nothing of any value in helping them at all.

They just don't get the vaguely hostile `n/a` string to tell them we don't
have anything helpful to say.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>