Factor out complicated slice dispatch equality logic into a method
This lifts the rich, complex comparison behaviour in slice command dispatching
out of the (four line) condition part of an if statement into a private top
level method.

There is no behavioural change, but the new code is substantially easier to
follow, and subsumes one additional helper method that lifted some of the
logic out of the conditional.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>