Use the stack, not member variables, to track command dispatch state
This rewrites part of the command dispatch code to track state - the current
position in the dispatch tree - using the stack rather than a mutable
member variable.

Previously multiple functions recursively called each other, mutating the
member variable tracking state as they did so.  This made for some pretty hard
to follow code when it came to determining who and what owned that data.

This replaces that with a simpler paradigm, passing the current state as an
argument to the function.  That makes it much clearer where the data is at
each level, and where it came from.

This does not completely eliminate the use of mutable state variables in the
dispatch code - which still tracks the arguments from the user in a member -
but it does cut the tracked state in half.

There are not externally visible behavioural differences from this change.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>