Use `get_node_instance_info` to target stopping node processes
Previously, the Razor daemon tried to shut down the API instances using
`killall` on anything named `node` - which failed on two counts, being that it
would kill too many things, and that it didn't work on some distributions
which ended up naming the process `nodejs` instead.

This replaces that with a more targeted process that kills only the nodejs
processes directly associated with Razor.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>