Turn YAML option data into Ruby code option data
This imports the YAML option definitions into Ruby code, avoiding the need to
parse and decode an external file before we can print option data.  This also
makes it harder - though, obviously, still entirely possible - to miss the
existence of that data when updating the code.

There is no behavioural change in the product from this.  The same data, in
the same form and order, is supplied to the rest of the system.  Only the
source is modified.

This closes #459, as the data now lives in Ruby.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>