Fix mismatched option `invert` with code `inverse`
An earlier fix caught most, but not all, of the mismatching uses of `invert`
and `inverse` through the code.  Jan Meyer <jan.meyer@m-bs.de> correctly noted
one of the places that was missed was the external option data files.

This, in part, is the motivation for pulling that data into the Ruby code -
having established that it is harder to discover that in a separate file, when
you search for a string, I wanted to tackle the root cause.

Ultimately, though, I need to apply the fix that Jan supplied - and having
rewritten the code this applies the concept rather than the delta he offered.

This supplants #451, which made the change in the YAML files.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>