(#464) Facter should only be installed if necessary in os_boot.erb
This change adds the option --enablerepo=base to the yum install call
in os_boot.erb to make sure the base repo is always enabled. This is
necessary because the factor package from the puppetlabs repo depends
on other packages like virt-what for example which are only available
in the base repo.

The change should not introduce any regressions because it simply
makes sure that the base repo is enabled. If it is enabled already,
this will not make any difference.

This pull request fixes issue #464.