(#471) Inconsistent assumption about @node_ip in models
This change fixes the broker_agent_handoff method in the models that
potentially used an empty ipaddress option when calling the agent_hand_off
method in the broker. Now, the agent_hand_off method is only called when
the @node_ip is not empty and still reports a useful error message when
it is empty.

Closes #471