(#478) /etc/rc.local should remain a symlink on Red Hat
This change adds the --follow-symlinks option to the sed command in the
os_complete.erb template of the redhat model. This is necessary because
the file /etc/rc.local on Red Hat should be a symlink to rc.d/rc.local
but, without this option, it becomes a regular file.

Note that the same change should not be necessary in Ubuntu and Debian
models because /etc/rc.local is not a symlink on those operating systems.

Closes #478