Update CI testing Puppet installation script repo setup
Puppet repositories have recently changed, resulting in a need to update the
scripts we use in the CI system to reflect that.  This allows Puppet to be
installed as part of the test regime.

This has not changed the broker installation process in any way, since that is
not using any of the repositories that have changed structure.

There are not functional changes to Razor as a result of this work, as it is
strictly related to Puppet Labs internal CI infrastructure.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>